Added zarr extraction and metadata
Two sections appended to the notebook which add extraction to a zarr array, and the ability to save metadata about the trench zarr to json files. These serve as a record and can be loaded back into the pipeline so that information about FOVs, timepoints and channels is not lost when extracting the trench zarr.